+++ src/lib/model/code.ts
@@ -1,9 +1,9 @@
 import { useChainId, useContractStore } from "lib/hooks";
+import { useCodeInfoByCodeId } from "lib/services/codeService";
 import {
-  useCodeInfoByCodeId,
-  useContractListByCodeId,
+  useContractListByCodeIdPagination,
   useContractListCountByCodeId,
-} from "lib/services/codeService";
+} from "lib/services/contractService";
 import {
   usePublicProjectByCodeId,
   usePublicProjectBySlug,
@@ -52,7 +52,7 @@ export const useCodeContractInstances = (
   offset: number,
   pageSize: number
 ): ContractInstances => {
-  const { data: contractList } = useContractListByCodeId(
+  const { data: contractList } = useContractListByCodeIdPagination(
     codeId,
     offset,
     pageSize