+++ src/lib/services/publicProjectService.ts
@@ -13,6 +13,8 @@ import type {
   PublicProjectInfo,
   RawPublicContract,
   RawPublicProjectInfo,
+  PublicCode,
+  RawPublicCode,
 } from "lib/types";
 
 const parseContract = (raw: RawPublicContract): PublicContract => ({
@@ -23,6 +25,11 @@ const parseContract = (raw: RawPublicContract): PublicContract => ({
   label: raw.label,
   instantiator: raw.instantiator,
   admin: raw.admin,
+});
+
+const parseCode = (raw: RawPublicCode): PublicCode => ({
+  ...raw,
+  contractCount: raw.contracts,
 });
 
 export const usePublicProjects = () => {
@@ -41,6 +48,7 @@ export const usePublicProjects = () => {
       .then(({ data: projects }) =>
         projects.map<PublicProjectInfo>((project) => ({
           ...project,
+          codes: project.codes.map(parseCode),
           contracts: project.contracts.map(parseContract),
         }))
       );
@@ -58,16 +66,20 @@ export const usePublicProjectBySlug = (slug: Option<string>) => {
     if (!slug) throw new Error("No project selected (usePublicProjectBySlug)");
     if (!currentChainRecord)
       throw new Error("No chain selected (usePublicProjectBySlug)");
-    return axios
-      .get<RawPublicProjectInfo>(
-        `${CELATONE_API_ENDPOINT}/projects/${getChainApiPath(
-          currentChainRecord.chain.chain_name
-        )}/${getMainnetApiPath(currentChainRecord.chain.chain_id)}/${slug}`
-      )
-      .then<PublicProjectInfo>(({ data: project }) => ({
-        ...project,
-        contracts: project.contracts.map(parseContract),
-      }));
+    return (
+      axios
+        .get<RawPublicProjectInfo>(
+          `${CELATONE_API_ENDPOINT}/projects/${getChainApiPath(
+            currentChainRecord.chain.chain_name
+          )}/${getMainnetApiPath(currentChainRecord.chain.chain_id)}/${slug}`
+        )
+        // eslint-disable-next-line sonarjs/no-identical-functions
+        .then<PublicProjectInfo>(({ data: project }) => ({
+          ...project,
+          codes: project.codes.map(parseCode),
+          contracts: project.contracts.map(parseContract),
+        }))
+    );
   }, [currentChainRecord, slug]);
 
   return useQuery(