+++ src/lib/types/contract.ts
@@ -43,36 +43,7 @@ export interface ContractMigrationHistory {
   height: number;
   timestamp: Date;
   remark: ContractHistoryRemark;
-}
-
-export enum ProposalStatus {
-  DEPOSIT_PERIOD = "DepositPeriod",
-  VOTING_PERIOD = "VotingPeriod",
-  PASSED = "Passed",
-  REJECTED = "Rejected",
-  FAILED = "Failed",
-  INACTIVE = "Inactive",
-}
-
-export enum ProposalType {
-  STORE_CODE = "StoreCode",
-  INSTANTIATE_CONTRACT = "InstantiateContract",
-  MIGRATE_CONTRACT = "MigrateContract",
-  UPDATE_ADMIN = "UpdateAdmin",
-  CLEAR_ADMIN = "ClearAdmin",
-  EXECUTE_CONTRACT = "ExecuteContract",
-  SUDO_CONTRACT = "SudoContract",
-}
-
-export interface ContractRelatedProposals {
-  proposalId: number;
-  title: string;
-  status: ProposalStatus;
-  votingEndTime: Date;
-  depositEndTime: Date;
-  resolvedHeight: Option<number | null>;
-  type: ProposalType;
-  proposer: Option<Addr>;
+  uploader: Addr;
 }
 
 export interface ContractData {